feat: add error handling for upsertUser, updateProduct, deleteProduct, and deleteComment functions

+++ backend/src/db/queries.ts
@@ -37,12 +37,22 @@ export const updateUser = async (id: string, data: Partial<NewUser>) => {
 
 // upsert => if user exists, update it, otherwise create a new one
 export const upsertUser = async (data: NewUser) => {
-  const existingUser = await getUserById(data.id);
-  if (existingUser) {
-    return updateUser(data.id, data);
-  }
+  // const existingUser = await getUserById(data.id);
+  // if (existingUser) {
+  //   return updateUser(data.id, data);
+  // }
 
-  return createUser(data);
+  // return createUser(data);
+
+  const [user] = await db
+    .insert(users)
+    .values(data)
+    .onConflictDoUpdate({
+      target: users.id,
+      set: data,
+    })
+    .returning();
+  return user;
 };
 
 // PRODUCT QUERY
@@ -81,6 +91,10 @@ export const getProductsByUserId = async (userId: string) => {
 };
 
 export const updateProduct = async (id: string, data: Partial<NewProduct>) => {
+  const existingProduct = await getProductById(id);
+  if (!existingProduct) {
+    throw new Error("Product not found");
+  }
   const [product] = await db
     .update(products)
     .set(data)
@@ -90,6 +104,10 @@ export const updateProduct = async (id: string, data: Partial<NewProduct>) => {
 };
 
 export const deleteProduct = async (id: string) => {
+  const existingProduct = await getProductById(id);
+  if (!existingProduct) {
+    throw new Error("Product not found");
+  }
   await db.delete(products).where(eq(products.id, id));
 };
 
@@ -101,6 +119,11 @@ export const createComment = async (data: NewComment) => {
 };
 
 export const deleteComment = async (id: string) => {
+  const existingComment = await getCommentsById(id);
+  if (!existingComment) {
+    throw new Error("Comment not found");
+  }
+
   const [comment] = await db
     .delete(comments)
     .where(eq(comments.id, id))
